twitter_bookmark_manager/deployment/final/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
sentence-transformers==2.3.1
qdrant-client==1.7.3
python-jose==3.3.0
gunicorn==21.2.0 